Minor UI Update, Changes to Saving Loading
Changed Yield and Animation Speed to use sliders.
Saving your project now prompts you to select a directory to save into.
Saving now generates a .block file instead of a json. to clarify to the user they need to export to use the block in game.
Loading now expects a .block file.